aidemy-bootstrap/assignment/requirements.txt
google-genai==1.0.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.28.0
google-cloud-aiplatform==1.76.0
functions-framework==3.8.2
langgraph==0.2.70
langchain_ollama==0.2.3
